[ALPS09798970] WIFI PCIe state duration update to UTC
[Description]
1. PCIe state duration update from tick to UTC:
  - update_time: FW update time in us
  - l0_time: state l0 time in us
  - l1_time: state l1 time in us
  - l1p2_time: state L1ss time in us

2. Add utility function convert_to_utc_usec in bridge
   to handle the utc conversion.

[Test]
Functional test and MBTF

MTK-Commit-Id: 3fb9d60865fce586157a4793a46d1845e41a6619

CR-Id: ALPS09798970
Feature: MBrain

+++ drivers/misc/mediatek/mbraink/mbraink_ioctl_struct_def.h
@@ -626,11 +626,11 @@ struct mbraink_dvfsrc_struct_data {
 
 struct mbraink_wifi2mbr_pcie_struct {
 	u64 timestamp;
-	unsigned int update_time;
+	u64 update_time;
 	unsigned int req_recovery_count;
-	unsigned int l0_time;
-	unsigned int l1_time;
-	unsigned int l1p2_time;
+	u64 l0_time;
+	u64 l1_time;
+	u64 l1p2_time;
 };
 
 struct mbraink_wifi2mbr_pcie_data {
